+++ src/test/java/com/vaadin/ui/formlayout/tests/FormLayoutIT.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.formlayout;
+package com.vaadin.ui.formlayout.tests;
 
 import java.util.List;
 
@@ -25,6 +25,7 @@ import org.openqa.selenium.WebElement;
 
 import com.vaadin.flow.demo.ComponentDemoTest;
 import com.vaadin.testbench.By;
+import com.vaadin.ui.formlayout.demo.FormLayoutView;
 
 /**
  * Integration tests for the {@link FormLayoutView}.
